Move replacment code to isolated module
